+++ Sources/TLS/AppleClient.swift
@@ -4,8 +4,10 @@
     import Foundation
     import Dispatch
 
+    /// An SSL client. Can be initialized by upgrading an existing socket or by starting an SSL socket.
     public final class AppleSSLClient: AppleSSLSocket {
-        public func initialize(hostname: String, certificate: Data? = nil) throws {
+        /// Upgrades the connection to SSL.
+        public func initialize(hostname: String, certificate: Certificate? = nil) throws {
             let context = try self.initialize(side: .clientSide)
             
             var hostname = [Int8](hostname.utf8.map { Int8($0) })
