+++ Sources/TLS/AppleServer.swift
@@ -2,8 +2,10 @@
     import Security
     import Foundation
     
-    public final class AppleSSLServer: AppleSSLSocket {
-        public func initialize(certificate: Data) throws {
+    /// A remote peer. This class should be used on a TCP server to accept SSL connections/upgrades.
+    public final class AppleSSLPeer: AppleSSLSocket {
+        /// Upgrades the peer to SSL
+        public func initialize(certificate: Certificate) throws {
             let context = try self.initialize(side: .serverSide)
             
             try self.setCertificate(to: certificate, for: context)
