+++ Sources/TLS/Error.swift
@@ -1,0 +1,8 @@
+enum Error: Swift.Error {
+    case cannotCreateContext
+    case writeError
+    case contextAlreadyCreated
+    case noSSLContext
+    case sslError(Int32)
+    case invalidCertificate
+}
